Implement /shop command for item purchase (#35)
Players are able to purchase items in with the /shop command

Items include pokeball, milk tea, energy with different price and benefits.
Players can use buttons to purchase each kind of items.
If player's balance is not enough, the purchase will fail.
A pokeball costs 25 coins. Players must use pokeball to catch new Pokemon.
A milk tea item costs 10 coins, which add hp by 4 for a random Pokemon in the team.
A energy bar item costs 15 coins, which add hp by 5 for a selected Pokemon in the team.
Players need to choose in a dropdown menu in order to select which Pokemon to use the energy bar item.